# api/openapi.yaml
openapi: 3.1.0
info:
  title: Reticora CMDB API
  version: 1.0.0
  description: |
    Multi-tenant Configuration Management Database API for MSPs and ISPs.
    Endpoints require tenant-scoped authentication unless explicitly marked otherwise.
  contact:
    name: DataHub Chiemgau
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/webhooks:
    get:
      operationId: listWebhooks
      tags: [Webhooks]
      summary: List webhook subscriptions
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: event
          in: query
          schema:
            $ref: '#/components/schemas/WebhookEvent'
      responses:
        '200':
          description: Paginated list of webhook subscriptions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createWebhook
      tags: [Webhooks]
      summary: Create a webhook subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '201':
          description: Webhook subscription created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookSubscription'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/entitlements:
    get:
      operationId: listEntitlements
      tags: [Entitlements]
      summary: List tenant entitlements
      responses:
        '200':
          description: Entitlements for the current tenant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntitlementListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: grantEntitlement
      tags: [Entitlements]
      summary: Grant or update an entitlement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantEntitlementRequest'
      responses:
        '201':
          description: Entitlement granted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entitlement'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/assets:
    get:
      operationId: listAssets
      tags: [Assets]
      summary: List assets
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/ClientID'
      responses:
        '200':
          description: Paginated list of assets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createAsset
      tags: [Assets]
      summary: Create an asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssetRequest'
      responses:
        '201':
          description: Asset created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/assignments:
    get:
      operationId: listAssignments
      tags: [Assignments]
      summary: List assignments
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: status
          in: query
          schema:
            type: string
        - name: assigned_to
          in: query
          schema:
            type: string
        - name: asset_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of assignments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssignmentListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createAssignment
      tags: [Assignments]
      summary: Create an assignment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssignmentRequest'
      responses:
        '201':
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/documents:
    get:
      operationId: listDocuments
      tags: [Documents]
      summary: List documents
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: category
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of documents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createDocument
      tags: [Documents]
      summary: Create a document metadata record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
      responses:
        '201':
          description: Document created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/stocktakes:
    get:
      operationId: listStocktakes
      tags: [Stocktake]
      summary: List stocktakes
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: status
          in: query
          schema:
            type: string
        - name: scope
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of stocktakes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StocktakeListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createStocktake
      tags: [Stocktake]
      summary: Create a stocktake
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStocktakeRequest'
      responses:
        '201':
          description: Stocktake created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stocktake'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/tickets:
    get:
      operationId: listTickets
      tags: [Tickets]
      summary: List tickets
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortDir'
        - name: status
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: assignee_id
          in: query
          schema:
            type: string
        - name: team_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Paginated list of tickets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createTicket
      tags: [Tickets]
      summary: Create a ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
      responses:
        '201':
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/collectors:
    get:
      operationId: listCollectors
      tags: [Collectors]
      summary: List discovery collectors
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Paginated list of collectors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectorListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: registerCollector
      tags: [Collectors]
      summary: Register a discovery collector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterCollectorRequest'
      responses:
        '201':
          description: Collector registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Collector'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/credentials:
    get:
      operationId: listCredentials
      tags: [Credentials]
      summary: List credential metadata
      description: >-
        Returns credential metadata (never the secret material) for the
        authenticated tenant. The organization is derived from the session and
        cannot be supplied by the client.
      responses:
        '200':
          description: Credential metadata list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Credential'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      operationId: createCredential
      tags: [Credentials]
      summary: Create an encrypted credential
      description: >-
        Stores a credential whose secret is envelope-encrypted at rest. The
        owning organization is derived from the authenticated tenant; any
        `organization_id` in the request body is ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCredentialRequest'
      responses:
        '201':
          description: Credential created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credential'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/forms:
    get:
      operationId: listFormDefinitions
      tags: [Forms]
      summary: List form definitions
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: client_id
          in: query
          schema: { type: string }
        - name: active
          in: query
          schema: { type: boolean }
      responses:
        '200': { description: Paginated form definitions, content: { application/json: { schema: { $ref: '#/components/schemas/FormDefinitionListResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      operationId: createFormDefinition
      tags: [Forms]
      summary: Create a form definition
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/CreateFormDefinitionRequest' } } }
      responses:
        '201': { description: Form definition created, content: { application/json: { schema: { $ref: '#/components/schemas/FormDefinition' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
  /api/v1/form-submissions:
    get:
      operationId: listAllFormSubmissions
      tags: [Forms]
      summary: List form submissions
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: form_id
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: ticket_id
          in: query
          schema: { type: string }
        - name: ci_id
          in: query
          schema: { type: string }
      responses:
        '200': { description: Paginated form submissions, content: { application/json: { schema: { $ref: '#/components/schemas/FormSubmissionListResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
  /api/v1/workflows:
    get:
      operationId: listWorkflows
      tags: [Workflows]
      summary: List workflow definitions
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: active
          in: query
          schema: { type: boolean }
      responses:
        '200': { description: Paginated workflows, content: { application/json: { schema: { $ref: '#/components/schemas/WorkflowDefinitionListResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
    post:
      operationId: createWorkflow
      tags: [Workflows]
      summary: Create a workflow definition
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateWorkflowDefinitionRequest' } } } }
      responses:
        '201': { description: Workflow created, content: { application/json: { schema: { $ref: '#/components/schemas/WorkflowDefinition' } } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
  /api/v1/workflow-runs:
    get:
      operationId: listWorkflowRuns
      tags: [Workflows]
      summary: List workflow runs
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: workflow_id
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200': { description: Paginated workflow runs, content: { application/json: { schema: { $ref: '#/components/schemas/WorkflowRunListResponse' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalServerError' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
      description: Maximum number of items to return.
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: |
        Number of items to skip before returning data. Ignored when `cursor` is
        supplied.
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: |
        Opaque keyset cursor taken from `next_cursor` of the previous page.
        Cursors are bound to the sort order they were issued for; sending one
        with a different `sort_by`/`sort_dir` combination yields 400.
    Search:
      name: search
      in: query
      schema:
        type: string
      description: Free-text search term.
    SortBy:
      name: sort_by
      in: query
      schema:
        type: string
      description: Sort field name.
    SortDir:
      name: sort_dir
      in: query
      schema:
        type: string
        enum: [asc, desc]
      description: Sort direction.
...
    ClientID:
      name: client_id
      in: query
      schema:
        type: string
  responses:
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    InternalServerError:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
...
    WebhookEvent:
      type: string
      enum: [ci.created, ci.updated, ci.deleted, ci.status_changed, relationship.created, relationship.deleted, discovery.completed]
    WebhookSubscription:
      type: object
      required: [id, organization_id, name, url, events, is_active, created_at, updated_at]
      properties:
        id:
          type: string
        organization_id:
          type: string
        name:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEvent'
        is_active:
          type: boolean
        headers:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateWebhookRequest:
      type: object
      required: [name, url, secret, events]
      properties:
        name:
          type: string
        url:
          type: string
          format: uri
        secret:
          type: string
          writeOnly: true
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEvent'
        headers:
          type: object
          additionalProperties:
            type: string
...
    WebhookListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/WebhookSubscription'
...
    Entitlement:
      type: object
      required: [organization_id, feature_key, plan, enabled]
      properties:
        organization_id:
          type: string
        feature_key:
          type: string
        plan:
          $ref: '#/components/schemas/EntitlementPlan'
        limit:
          type: integer
          description: Maximum number of records for the feature. 0 means unlimited.
        enabled:
          type: boolean
        expires_at:
          type: string
          format: date-time
...
    GrantEntitlementRequest:
      type: object
      required: [feature_key]
      properties:
        feature_key:
          type: string
        plan:
          $ref: '#/components/schemas/EntitlementPlan'
        enabled:
          type: boolean
        limit:
          type: integer
        expires_at:
          type: string
          format: date-time
    EntitlementListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Entitlement'
    Asset:
      type: object
      required: [id, organization_id, asset_tag, name, category, status, custom_fields, created_at, updated_at]
      properties:
        id:
          type: string
        organization_id:
          type: string
        client_id:
          type: string
        ci_id:
          type: string
        asset_tag:
          type: string
        name:
          type: string
        category:
          type: string
        status:
          type: string
        purchase_date:
          type: string
          format: date
        purchase_cost:
          type: number
        currency:
          type: string
        warranty_end:
          type: string
          format: date
        supplier:
          type: string
        invoice_number:
          type: string
        serial_number:
          type: string
        location:
          type: string
        notes:
          type: string
        custom_fields:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateAssetRequest:
      type: object
      required: [asset_tag, name]
      properties:
        client_id:
          type: string
        ci_id:
          type: string
        asset_tag:
          type: string
        name:
          type: string
        category:
          type: string
        status:
          type: string
        purchase_date:
          type: string
          format: date
        purchase_cost:
          type: number
        currency:
          type: string
        warranty_end:
          type: string
          format: date
        supplier:
          type: string
        invoice_number:
          type: string
        serial_number:
          type: string
        location:
          type: string
        notes:
          type: string
        custom_fields:
          type: object
          additionalProperties: true
...
    AssetListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Asset'
    Assignment:
      type: object
      required: [id, organization_id, assigned_to, assigned_by, assignment_type, status, assigned_at, created_at, updated_at]
      properties:
        id:
          type: string
        organization_id:
          type: string
        asset_id:
          type: string
        ci_id:
          type: string
        assigned_to:
          type: string
        assigned_by:
          type: string
        assignment_type:
          type: string
        status:
          type: string
        assigned_at:
          type: string
          format: date-time
        due_date:
          type: string
          format: date-time
        returned_at:
          type: string
          format: date-time
        return_condition:
          type: string
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateAssignmentRequest:
      type: object
      required: [assigned_to]
      properties:
        asset_id:
          type: string
        ci_id:
          type: string
        assigned_to:
          type: string
        assignment_type:
          type: string
        due_date:
          type: string
          format: date-time
        notes:
          type: string
    AssignmentListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Assignment'
    Document:
      type: object
      required: [id, organization_id, title, file_name, file_size, mime_type, storage_key, version, category, tags, uploaded_by, created_at, updated_at]
      properties:
        id:
          type: string
        organization_id:
          type: string
        title:
          type: string
        description:
          type: string
        file_name:
          type: string
        file_size:
          type: integer
        mime_type:
          type: string
        storage_key:
          type: string
        version:
          type: integer
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        uploaded_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateDocumentRequest:
      type: object
      required: [title, file_name, storage_key]
      properties:
        title:
          type: string
        description:
          type: string
        file_name:
          type: string
        file_size:
          type: integer
        mime_type:
          type: string
        storage_key:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
...
    DocumentListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Document'
    Stocktake:
      type: object
      required: [id, organization_id, title, status, scope, total_expected, total_scanned, total_missing, total_surplus, created_at, updated_at]
      properties:
        id:
          type: string
        organization_id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
        scope:
          type: string
        started_by:
          type: string
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        due_date:
          type: string
          format: date-time
        total_expected:
          type: integer
        total_scanned:
          type: integer
        total_missing:
          type: integer
        total_surplus:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateStocktakeRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        scope:
          type: string
        due_date:
          type: string
          format: date-time
        total_expected:
          type: integer
...
    StocktakeListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Stocktake'
...
    Ticket:
      type: object
      required: [id, organization_id, ticket_number, title, status, priority, category, reporter_id, tags, created_at, updated_at]
      properties:
        id:
          type: string
        organization_id:
          type: string
        ticket_number:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        category:
          type: string
        reporter_id:
          type: string
        assignee_id:
          type: string
        team_id:
          type: string
        related_ci_id:
          type: string
        related_asset_id:
          type: string
        due_date:
          type: string
          format: date-time
        resolved_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateTicketRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        category:
          type: string
        assignee_id:
          type: string
        team_id:
          type: string
        related_ci_id:
          type: string
        related_asset_id:
          type: string
        due_date:
          type: string
          format: date-time
        tags:
          type: array
          items:
            type: string
...
    TicketListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Ticket'
...
    CollectorListResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationEnvelope'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Collector'
...